requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
crispy-forms-gds==0.2.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
djangorestframework==3.12.1
filelock==3.0.12
gunicorn==20.0.4
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.21
pbr==5.5.0
Pillow==7.2.0
psycopg2==2.8.6
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
stevedore==3.2.2
toml==0.10.1
virtualenv==20.0.31
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
wrapt==1.12.1
